chore: 优化构建流程

+++ build.sh
@@ -1,10 +1,15 @@
 #!/bin/sh
 
-if [ -d "./standalone" ]; then
-  rm -rf ./standalone
+if [ -d "./build" ]; then
+  rm -rf ./build
 fi
 
 cp -r public .next/standalone/ 
 cp -r .next/static .next/standalone/.next/ 
-cp pm2.json .next/standalone/
-mv .next/standalone ./
+
+if [ -f pm2.json ]; then
+  cp pm2.json .next/standalone/
+fi
+
+mv .next/standalone ./build
+rm -rf ./.next
